xmlify_json and jsonify_xml
Adds base methods to the glovebox program to parse valid XML JSON into XML strings as well as convert XML into a JSON.
